backend/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-jose
pydantic
pymongo
motor
pydantic-settings==2.2.1